+++ src/lib/components/Accordion/types.ts
@@ -1,10 +1,19 @@
 import type { Props, Handler, Transition } from '$internal';
 
 export interface AccordionProps extends Props<HTMLDivElement> {
-	/** The value of the active item. */
-	value?: string;
+	/**
+	 * The value of the active item.\
+	 *
+	 * If items have no value set, the unique ID will be used.
+	 */
+	value?: string[];
 	/** Allow only a singluar `AccordionItem` to be opened at once. */
 	single?: boolean;
+	/**
+	 * An event that fires everytime the Accordion `value` prop changes.
+	 * @param values The current values of the Accordion.
+	 */
+	onChange?: (values: string[]) => void;
 }
 
 export interface AccordionContentProps extends Props<HTMLDivElement> {
@@ -45,6 +54,12 @@ interface AccordionItemState {
 	disabled: boolean;
 }
 export interface AccordionItemProps extends Props<HTMLDivElement, AccordionItemState> {
+	/**
+	 * The value of the accordion item.\
+	 *
+	 * If no value is set, unique ID is used.
+	 */
+	value?: string;
 	/** Disables the items, disallowing clicking and keyboard navigation. */
 	disabled?: boolean;
 }